Milestone 4/backend/requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
greenlet==3.1.1

# Pydantic
pydantic==2.9.2
pydantic-settings==2.6.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# GenAI - Knowledge Assistant
google-generativeai==0.4.1
chromadb==0.5.20
sentence-transformers==3.3.1
PyPDF2==3.0.1

# Study Guide Generator - YouTube
youtube-transcript-api==0.6.2
pytube==15.0.0

# Async file handling
aiofiles==24.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

PyPDF2==3.0.1
google-generativeai==0.4.1
python-pptx==0.6.23    # Optional for PowerPoint export

langchain==0.1.20
langchain-google-genai==0.0.11
langchain-community==0.0.38

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.2
